BZ #1585272 - CV Publish: copy correct errata
This commit makes the following changes to the cv publish process
Before This Commit:
Errata would get copied over to the destination repo if ANY of the
packages belonging to the errata were copied.

After this Commit:
Errata would get copied over to the destination repo only if ALL
of the packages belonging to the errata were copied.

This commit also includes changes to hide modular errata in a content
view filter (since all modules and module rpms are copied over, all the
modular errata will also get copied automatically.)