Fixes #31636 - support corrupted/missing content workflow
as part of the pulp3 migration.

This adds knowldge around if a content unit is corrupted/missing and
whether the user has marked it for approval or not.  It also adds rake
tasks for approving/unapproving content and updates migration/switchover
appropriately.  Only Rpms/Files can be put through this process,
corrupted/missing docker content will just be ignored on upgrade.